reconstruction v2

+++ app/src/main/java/com/impact/thebestweather/models/weather/daily/Minimum.kt
@@ -3,5 +3,5 @@ package com.impact.thebestweather.models.weather.daily
 data class Minimum(
     val Unit: String,
     val UnitType: Int,
-    val Value: Int
+    val Value: Double
 )
